+++ src/main/java/com/eureka/spartaonetoone/domain/user/application/dtos/request/UserUpdateRequestDto.java
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class UserUpdateRequestDto {
 
-    @NotBlank(message = "사용자 이름은 필수입니다.")
-    private String username;
+	@NotBlank(message = "사용자 이름은 필수입니다.")
+	private String username;
 
-    @NotBlank(message = "이메일은 필수입니다.")
-    private String email;
+	@NotBlank(message = "이메일은 필수입니다.")
+	private String email;
 }
